+++ settfex/__init__.py
@@ -35,7 +35,7 @@ Usage:
     >>> asyncio.run(main())
 """
 
-__version__ = "0.13.0"
+__version__ = "0.14.0"
 __author__ = "batt"
 __license__ = "MIT"